packages/query-registry/pyproject.toml
[project]
dependencies = [
    "attrs>=25.3.0",
    "snowflake-connector-python>=3.0.0",
]
description = "Query registry for async execute_query management"
name = "query-registry"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[build-system]
build-backend = "uv_build"
requires = [
    "uv_build>=0.8.8,<0.9.0",
]
